fix(test): eliminate mock pollution flake in getEagerServiceAvailability
Extract getEagerServiceAvailability to service-availability.ts — a pure
utility with zero React imports. The test now imports it directly without
mock.module(), eliminating cross-file mock pollution.

Root cause: TunnelPanel.test.tsx mocks @/hooks/useServiceChannel with a
hardcoded `getEagerServiceAvailability: () => false` stub. When both
files ran in the same Bun worker, the dynamic import in the old
useServiceChannel.test.ts could resolve to the mock instead of the real
module, causing the 'returns true' test to fail intermittently.

useServiceChannel.ts re-exports from service-availability.ts so existing
consumers are unaffected.
